refactor: isolate tag team suspension validation
