+++ app/Models/TagTeams/TagTeam.php
@@ -16,6 +16,7 @@ use App\Models\Concerns\IsSuspendable;
 use App\Models\Concerns\ProvidesTagTeamWrestlers;
 use App\Models\Concerns\ValidatesTagTeamEmployment;
 use App\Models\Concerns\ValidatesTagTeamLifecycle;
+use App\Models\Concerns\ValidatesTagTeamSuspension;
 use App\Models\Contracts\Bookable;
 use App\Models\Contracts\CanBeAStableMember;
 use App\Models\Contracts\CanBeChampion;
@@ -151,6 +152,7 @@ class TagTeam extends Model implements Bookable, CanBeAStableMember, CanBeChampi
     use SoftDeletes;
     use ValidatesTagTeamEmployment;
     use ValidatesTagTeamLifecycle;
+    use ValidatesTagTeamSuspension;
 
     /**
      * The number of the wrestlers allowed on a tag team.
