+++ docs/architecture/lifecycle-operation-boundaries.md
@@ -153,6 +153,8 @@ Tag-team lifecycle validation uses the concrete `Wrestler` models returned by it
 
 Tag-team employment eligibility is isolated in `ValidatesTagTeamEmployment`. The concern owns only the model-level `canBeEmployed()` and `ensureCanBeEmployed()` rules; the employment Action continues to own orchestration and persistence. Other tag-team lifecycle dimensions remain separate and will be extracted independently.
 
+Tag-team suspension eligibility is isolated in `ValidatesTagTeamSuspension`. Suspension and reinstatement remain paired as opposite transitions of one lifecycle dimension, while their Actions continue to own suspension-period persistence, dates, transactions, and member cascades.
+
 The following generalized classes were confirmed to have no production, configuration, container, console, route, or test consumers and were removed rather than retained as foundations for the target architecture:
 
 - `ActionPipeline`;
